# back-end/public/api/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Knowledge Learning API",
        "description": "Documentation OpenAPI générée depuis le code (Symfony 8 / PHP 8.4)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/verify-email": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Vérifier un compte via URL signée",
                "description": "Handles email verification via a signed URL.",
                "operationId": "a00843a2164e95f5ff03a17d358f813d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 123
                    },
                    {
                        "name": "signature",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "expires",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Compte vérifié"
                    },
                    "400": {
                        "description": "Lien invalide ou expiré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lien invalide ou expiré."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur introuvable."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/payments/checkout": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Créer une session Stripe Checkout (cursus/leçon)",
                "description": "Handle checkout session creation for cursus or lesson purchase.",
                "operationId": "2d95e49e63dc12eb48752086eb519867",
                "parameters": [
                    {
                        "name": "X-CSRF-TOKEN",
                        "in": "header",
                        "description": "Token CSRF requis (session cookie). CSRF id: auth",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "itemId"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "cursus",
                                            "lesson"
                                        ],
                                        "example": "cursus"
                                    },
                                    "itemId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "sessionId",
                                        "checkoutUrl"
                                    ],
                                    "properties": {
                                        "sessionId": {
                                            "type": "string",
                                            "example": "cs_test_123"
                                        },
                                        "checkoutUrl": {
                                            "type": "string",
                                            "example": "https://checkout.stripe.com/..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Refusé (CSRF invalide ou item non achetable)"
                    },
                    "404": {
                        "description": "Item introuvable"
                    },
                    "409": {
                        "description": "Déjà acheté"
                    },
                    "422": {
                        "description": "Payload invalide"
                    },
                    "502": {
                        "description": "Erreur Stripe"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}